+++ src/Traits/EncoderAware.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace PhpXmlRpc\JsonRpc\Traits;
+
+use PhpXmlRpc\JsonRpc\Encoder;
+
+trait EncoderAware
+{
+    protected static $encoder;
+
+    public function getEncoder()
+    {
+        if (self::$encoder === null) {
+            self::$encoder = new Encoder();
+        }
+        return self::$encoder;
+    }
+
+    public static function setEncoder($encoder)
+    {
+        self::$encoder = $encoder;
+    }
+}