+++ src/Value.php
@@ -2,28 +2,12 @@
 
 namespace PhpXmlRpc\JsonRpc;
 
-use PhpXmlRpc\JsonRpc\Helper\Serializer;
+use PhpXmlRpc\JsonRpc\Traits\SerializerAware;
 use PhpXmlRpc\Value as BaseValue;
 
-/**
- * @todo once we make php 5.4 a mandatory requirement, implement a SerializerAware trait
- */
 class Value extends BaseValue
 {
-    protected static $serializer;
-
-    public function getSerializer()
-    {
-        if (self::$serializer === null) {
-            self::$serializer = new Serializer();
-        }
-        return self::$serializer;
-    }
-
-    public static function setSerializer($serializer)
-    {
-        self::$serializer = $serializer;
-    }
+    use SerializerAware;
 
     /**
      * Returns json representation of the value.